refactor class and function module
