+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/Keyboard.iml" filepath="$PROJECT_DIR$/.idea/Keyboard.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/keyboard.iml" filepath="$PROJECT_DIR$/.idea/keyboard.iml" />
     </modules>
   </component>
 </project>
